ログイン追加

+++ app/Http/Controllers/bbController.php
@@ -6,29 +6,52 @@ use Illuminate\Http\Request;
 use App\Http\Controllers\Controller;
 use App\bb;
 use App\Http\Requests\bbRequest;
-
+use Illuminate\Support\Facades\Auth;
 class bbController extends Controller
 {
+    /**
+     * 初期
+     * @access public
+     * @param bbRequest $request
+     */
     public function index()
-    {
+    {        
+        if (!Auth::check()) {
+            return view('home');
+        }        
         // DBよりBookテーブルの値を全て取得
         $bbs = bb::all();
-
         // 取得した値をビュー「book/index」に渡す
         return view('bb/index', compact('bbs'));
+
     }
 
+    /**
+     * 編集
+     * @access public
+     * @param bbRequest $request
+     */
     public function edit($id)
     {
+        if (!Auth::check()) {            
+            return view('home');
+        }   
         // DBよりURIパラメータと同じIDを持つBookの情報を取得
         $bb = bb::findOrFail($id);
 
         // 取得した値をビュー「book/edit」に渡す
         return view('bb/edit', compact('bb'));
     }
-
+    /**
+     * 更新
+     * @access public
+     * @param bbRequest $request
+     */
     public function update(bbRequest $request, $id)
     {
+        if (!Auth::check()) {
+            return view('home');
+        }            
         $bb = bb::findOrFail($id);
         $bb->name = $request->name;
         $bb->price = $request->price;
@@ -37,22 +60,47 @@ class bbController extends Controller
 
         return redirect("/bb");
     }
+
+    /**
+     * 削除
+     * @access public
+     * @param bbRequest $request
+     */
     public function destroy($id)
     {
+        if (!Auth::check()) {
+            return view('home');
+        }            
         $bb = bb::findOrFail($id);
         $bb->delete();
 
         return redirect("/bb");
     }
+
+    /**
+     * 新規作成ボタン
+     * @access public
+     * @param bbRequest $request
+     */
     public function create()
     {
+        if (!Auth::check()) {
+            return view('home');
+        }            
         // 空の$bookを渡す
         $bb = new bb();
         return view('bb/create', compact('bb'));
     }
-
+    /**
+     * 新規作成処理
+     * @access public
+     * @param bbRequest $request
+     */
     public function store(bbRequest $request)
     {
+        if (!Auth::check()) {
+            return view('home');
+        }            
         $bb = new bb();
         $bb->name = $request->name;
         $bb->price = $request->price;
